requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Web & API
requests==2.31.0
httpx==0.25.1
websockets==12.0
python-socketio==5.10.0
flask==3.0.0
protobuf==4.25.0  # Required for market feed proto messages

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
joblib==1.3.2

# Deep Learning (PyTorch - CPU only for Docker)
torch==2.2.2
pytorch-forecasting==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Time-Series Database
timescale==0.1.0

# Cache
redis==5.0.1
hiredis==2.2.3

# Message Queue (Optional)
pika==1.3.2

# Financial Calculations
py_vollib==1.0.1  # Black-Scholes Greeks
# Technical Analysis
# pandas-ta>=0.3.14  # Not available in PyPI - commented out
# ta-lib==0.4.28  # Requires system library - commented out

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
apscheduler==3.10.4  # For EOD training scheduler

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
